+++ clang/lib/Serialization/ASTReaderStmt.cpp
@@ -1452,17 +1452,6 @@ void ASTStmtReader::VisitConvertVectorExpr(ConvertVectorExpr *E) {
         FPOptionsOverride::getFromOpaqueInt(Record.readInt()));
 }
 
-void ASTStmtReader::VisitConvertFromArbitraryFPExpr(
-    ConvertFromArbitraryFPExpr *E) {
-  VisitExpr(E);
-  E->BuiltinLoc = readSourceLocation();
-  E->RParenLoc = readSourceLocation();
-  E->TInfo = readTypeSourceInfo();
-  E->SrcExpr = Record.readSubExpr();
-  E->Format = static_cast<ConvertFromArbitraryFPExpr::ArbitraryFPFormat>(
-      Record.readInt());
-}
-
 void ASTStmtReader::VisitBlockExpr(BlockExpr *E) {
   VisitExpr(E);
   E->setBlockDecl(readDeclAs<BlockDecl>());
@@ -3535,10 +3524,6 @@ Stmt *ASTReader::ReadStmtFromStream(ModuleFile &F) {
       S = ConvertVectorExpr::CreateEmpty(Context, HasFPFeatures);
       break;
     }
-
-    case EXPR_CONVERT_FROM_ARBITRARY_FP:
-      S = new (Context) ConvertFromArbitraryFPExpr(Empty);
-      break;
 
     case EXPR_BLOCK:
       S = new (Context) BlockExpr(Empty);