+++ clang/tools/libclang/CIndex.cpp
@@ -2114,7 +2114,6 @@ public:
   void VisitDeclStmt(const DeclStmt *S);
   void VisitDependentScopeDeclRefExpr(const DependentScopeDeclRefExpr *E);
   void VisitDesignatedInitExpr(const DesignatedInitExpr *E);
-  void VisitConvertFromArbitraryFPExpr(const ConvertFromArbitraryFPExpr *E);
   void VisitExplicitCastExpr(const ExplicitCastExpr *E);
   void VisitForStmt(const ForStmt *FS);
   void VisitGotoStmt(const GotoStmt *GS);
@@ -3162,11 +3161,6 @@ void EnqueueVisitor::VisitDesignatedInitExpr(const DesignatedInitExpr *E) {
     AddStmt(E->getArrayRangeEnd(D));
     AddStmt(E->getArrayRangeStart(D));
   }
-}
-void EnqueueVisitor::VisitConvertFromArbitraryFPExpr(
-    const ConvertFromArbitraryFPExpr *E) {
-  EnqueueChildren(E);
-  AddTypeLoc(E->getTypeSourceInfo());
 }
 void EnqueueVisitor::VisitExplicitCastExpr(const ExplicitCastExpr *E) {
   EnqueueChildren(E);